Add check for .env.local and module bindings generations to only write on changes

+++ crates/cli/src/subcommands/generate.rs
@@ -200,7 +200,9 @@ pub async fn exec_ex(
             fs::create_dir_all(out_dir.join(parent))?;
         }
         let path = out_dir.join(fname);
-        fs::write(&path, code)?;
+        if !path.exists() || fs::read_to_string(&path)? != code {
+            fs::write(&path, code)?;
+        }
         paths.insert(path);
     }
